Refactor quiz results display
